Flytter kodeterm-funksjonene fra Utils til src/kodeverk/index.js
