+++ src/kodeverk/index.js
@@ -32,6 +32,26 @@ const termTilKode = (verdi, muligeValg) => {
   return valgtKodeverkObjekt && objektTilKode(valgtKodeverkObjekt);
 };
 
+function termFraKTObject(kodetermListe, kode) {
+  for (let i = 0, j = kodetermListe.length; i < j; i += 1) {
+    if (kodetermListe[i].kode === kode) {
+      return kodetermListe[i].term;
+    }
+  }
+  return null;
+}
+
+function termFraNestedKTObject(nestedKTObject, kode) {
+  const kodetermListe = [];
+  // eslint-disable-next-line no-restricted-syntax
+  for (const kodeterm in nestedKTObject) {
+    if (Object.prototype.hasOwnProperty.call(nestedKTObject, kodeterm)) {
+      kodetermListe.push(...nestedKTObject[kodeterm]);
+    }
+  }
+  return termFraKTObject(kodetermListe, kode);
+}
+
 export {
   AvsenderTyper,
   Koder,
@@ -44,5 +64,7 @@ export {
   kodeTilTerm,
   kodeTilObjekt,
   termTilKode,
+  termFraKTObject,
+  termFraNestedKTObject,
   Panel,
 };
